refactor:  for added privacy,  tweak the md5name
Given the need to not expose what files are actually being stored,
the md5name name/folder/s3key will be derived from a env var and
the md5 name.  Assuming the file is also encypted and compressed,
this should offer substantial masking of  what the file actually is.
However, the plan is still to expose the type in the path, but that
is subject to change upon further review.