change: add support for new volume type features
In order to show and store more info, and perhaps programatically
alter the way the programs based on the volume type,
this change adds both a prompt to the user and some
data storeage for the type of volume.  At the time of change,
the key need is to differentiate phyical from non-physical,
so that when the flask app goes to get the file, we can be
more clear about which onese are truely expected to be
available.  USB and VHD sources, may need to be remounted,
and are therefore less likely to be available.